Create Updatable concern
