+++ dashboard/test/models/concerns/user/updateable_test.rb
@@ -1,0 +1,5 @@
+require 'test_helper'
+
+class UpdateableTest < ActiveSupport::TestCase
+  include Minitest::RSpecMocks
+end
